protect delete item view from malicious get request
Nothing will happen if someone types the url that calls the view to delete the item in the browser and hits enter. The code will check if user is authenticated and if he/she is the owner of the list.
The list is obtained through the item, otherwise, the owner of list 1 could send "/lists/1/items/4/delete", and delete item 4, which doesn't necessarily belong to them, and the check would pass.